try increasing the timeouts for slow machines
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1161097 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/solrj/src/test/org/apache/solr/client/solrj/TestLBHttpSolrServer.java
@@ -162,8 +162,8 @@ public class TestLBHttpSolrServer extends LuceneTestCase {
     }
     HttpClient myHttpClient = new HttpClient(new MultiThreadedHttpConnectionManager());
 
-    myHttpClient.getParams().setParameter("http.connection.timeout", new Integer(100));
-    myHttpClient.getParams().setParameter("http.socket.timeout", new Integer(100));
+    myHttpClient.getParams().setParameter("http.connection.timeout", new Integer(250));
+    myHttpClient.getParams().setParameter("http.socket.timeout", new Integer(250));
     LBHttpSolrServer lbHttpSolrServer = new LBHttpSolrServer(myHttpClient, s);
     lbHttpSolrServer.setAliveCheckInterval(500);
